// src/interfaces/common.tsx
export interface Page<T> {
  content: T[];
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  first: boolean;
  last: boolean;
  numberOfElements: number;
  empty: boolean;
}

export interface ErrorsDTO {
  message: string;
  timestamp: string;
  listaErrori?: string[];
}